server-flask/requirements.txt
flask
flask-cors
pandas
kafka-python